requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.2
pipenv==2018.11.26
pkginfo==1.5.0.1
pluggy==0.13.0
py==1.8.0
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.2.2
python-dateutil==2.8.0
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
six==1.13.0
SQLAlchemy==1.3.11
tqdm==4.37.0
twine==2.0.0
urllib3==1.25.6
virtualenv==16.7.7
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
zipp==0.6.0
